add avatar & default avatar to user with migration & seeders

+++ src/modules/static-field/static-field.service.ts
@@ -3,12 +3,18 @@ import { InjectModel } from '@nestjs/sequelize'
 import { StaticField } from '../../database/models/singles/StaticField/static-field.model'
 import { BaseServiceRead } from '../../core/bases/services'
 import { CreateStaticFieldAttributes } from '../../database/models/singles/StaticField/static-field.attributes'
+import { ApiConfigService } from '../../core/modules/shared/services/api-config.service'
+import { unlink } from 'fs'
+import { join } from 'path'
+import { NotFoundException } from '../../core/exceptions/build-in'
+import { ErrorMessagesConstants } from '../../core/constants'
 
 @Injectable()
 export class StaticFieldService extends BaseServiceRead<StaticField> {
   constructor(
     @InjectModel(StaticField)
-    private readonly staticFieldRepository: typeof StaticField
+    private readonly staticFieldRepository: typeof StaticField,
+    private readonly configService: ApiConfigService
   ) {
     super({
       modelRepository: staticFieldRepository,
@@ -25,6 +31,20 @@ export class StaticFieldService extends BaseServiceRead<StaticField> {
   }
 
   async delete(id: number) {
-    return this.staticFieldRepository.destroy({ where: { id } })
+    const target = await this.staticFieldRepository.findOne({ where: { id } })
+    if (!target)
+      throw new NotFoundException(
+        ErrorMessagesConstants.NotFound,
+        'No such static-field'
+      )
+    this.deleteFromStorage(target.name)
+    return await this.staticFieldRepository.destroy({ where: { id } })
+  }
+
+  private deleteFromStorage(name: string) {
+    unlink(
+      join(this.configService.multerConfig.dest || './static/uploads', name),
+      console.error
+    )
   }
 }
